@@ -18,7 +18,8 @@ ActiveAdmin.register Child do
     :name_tag_first_name, :name_tag_last_name, :childcare_deposit,
     :childcare_comment, :rec_pass_start_at, :rec_pass_end_at, :forms_approved, :forms_approved_by,
     :tshirt_size,
-    childcare_weeks: [], hot_lunch_weeks: [], cost_adjustments_attributes: %i[
+    childcare_weeks: [], hot_lunch_weeks: [], 
+    cost_adjustments_attributes: %i[
       id _destroy description person_id price percent cost_type
     ],
     childcare_medical_history_attributes: %i[
