@@ -44,7 +44,7 @@ Rails.application.configure do
   # Force all access to the app over SSL, use Strict-Transport-Security, and use secure cookies.
   # if ENV["AWS_EXECUTION_ENV"].present?
   #   # Force all access to the app over SSL, use Strict-Transport-Security, and use secure cookies.
-  #   config.force_ssl = true
+  config.force_ssl = true
   #   config.ssl_options = {
   #     redirect: {
   #       exclude: -> (request) { request.fullpath == '/monitors/lb' }
