@@ -1,7 +1,12 @@
 require 'redis'
 require 'datadog/statsd'
 
-redis_conf = YAML.safe_load(ERB.new(File.read(Rails.root.join("config", "redis.yml"))).result, [Symbol], [], true)["sidekiq"]
+redis_conf = YAML.safe_load(
+  ERB.new(File.read(Rails.root.join("config", "redis.yml"))).result,
+  permitted_classes: [Symbol],
+  permitted_symbols: [],
+  aliases: true
+)["sidekiq"]
 
 redis_settings = {url: Redis.new(redis_conf).id}
 
