@@ -3,7 +3,9 @@ FactoryBot.define do
     email { Faker::Internet.email }
     role { User::ROLES.sample }
     guid { Faker::Number.hexadecimal(digits: 8) }
-
+    first_name { Faker::Name.first_name }
+    last_name  { Faker::Name.last_name }
+  
     factory :admin_user do
       role { 'admin' }
     end
